Pizzas/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
